requirements.txt
streamlit
opencv-python
mediapipe
faiss-cpu
numpy
pyttsx3
tensorflow
